Move `MembershipChange` to its own file.
