+++ libraries/matrix/api/src/main/kotlin/io/element/android/libraries/matrix/api/timeline/item/event/MembershipChange.kt
@@ -1,0 +1,37 @@
+/*
+ * Copyright (c) 2023 New Vector Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.element.android.libraries.matrix.api.timeline.item.event
+
+enum class MembershipChange {
+    NONE,
+    ERROR,
+    JOINED,
+    LEFT,
+    BANNED,
+    UNBANNED,
+    KICKED,
+    INVITED,
+    KICKED_AND_BANNED,
+    INVITATION_ACCEPTED,
+    INVITATION_REJECTED,
+    INVITATION_REVOKED,
+    KNOCKED,
+    KNOCK_ACCEPTED,
+    KNOCK_RETRACTED,
+    KNOCK_DENIED,
+    NOT_IMPLEMENTED;
+}
